:sparkles: add Redirects key - SHU-74
