+++ next.config.js
@@ -1,6 +1,21 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-}
 
-module.exports = nextConfig
+  async redirects() {
+    return [
+      {
+        source: '/((?!maintenance).*)',
+        destination: '/maintenance',
+        permanent: false,
+      },
+      // {
+      //   source: "/maintenance",
+      //   destination: "/",
+      //   permanent: false,
+      // },
+    ];
+  },
+};
+
+module.exports = nextConfig;
