+++ src/styles/Maintenance.module.scss
@@ -1,0 +1,12 @@
+.maintenanceContainer {
+  background-image: linear-gradient(316deg, #f093fb 0%, #f5576c 100%);
+  inset: 0;
+  padding-top: 15vh;
+  position: fixed;
+  text-align: center;
+  .text {
+    display: flex;
+    gap: 0.5em;
+    justify-content: center;
+  }
+}
